Changed how bodies and joints are initialized and added some safety checks
World no longer initialize bodies and joints on Start, but instead the NewtonBody component initialized itself on Start. The reason we let the world initialize the bodies was if the bodies would "start" before the world but since the world is now created already in it's constructor that shouldn't happen. Now instead the body and joint initialize themselves. Also added a safety check so that a body can't initialize more than once. That happened if you wrote a script that dynamically added bodies to the scene and called InitRigidBody. Then the body would be initialized twice, creating an extra native body and collider which would never be used. Also the body would register itself twice in the world list which caused nullreference exceptions. When the body is destroyed it unregister itself(once) but one reference would still be in the list.